+++ profileService/internal/domain/program/repository.go
@@ -4,5 +4,5 @@ import "context"
 
 type Repository interface {
 	GetAllPrograms(ctx context.Context) ([]*Program, error)
-	GetProgramNameByID(ctx context.Context, id int) (*string, error)
+	GetProgramNameByID(ctx context.Context, id int64) (*string, error)
 }
